Fixes #9: Validate file size before content type
Aside from the fact that checking size is cheaper, this change fixes
a misleading and confusing message described in issue #9, where a
user uploading a MS Word .doc file which is too big is told incorrectly
that the content type is wrong.

The tests were changed to work (only) with Django 1.6 and
above, which changed humanize to add a non-breaking space (\xa0)
between a value and its unit.  See this ticket for more
details:

  https://code.djangoproject.com/ticket/20246

I tested with these dependencies along with Python 2.7.6:

  Django==1.6.11
  python-magic==0.4.10